Possible medians

These functions are helpful if median2() returns NA:

median_possible_values() returns all values that might possibly be the median, given the number of missing values.

median_range() returns the minimal and maximal possible median values.

Usage

median_possible_values(x)

median_range(x)

# S3 method for default
median_possible_values(x)

# S3 method for default
median_range(x)

Arguments

x

Numeric or similar. Vector to search for its possible medians.

Value

Vector of the same type as x. Its values are unique and sorted by size. It always has length 2 when calling median_range().

If the median can be determined (i.e., median2() would return a non-NA

value), median_possible_values() returns just one value, and median_range() returns two identical values.

If there are missing values at the median position, the set of possible medians cannot be determined. The functions will then return NA or c(NA, NA), respectively.

Details

Like median2(), these functions are generic, so methods can be defined for other classes. This documentation describes the default methods.

Author

Lukas Jung, R Core Team

Examples

# All possible values,
# plus the range:
x1 <- c(7, 7, 7, 8, 9, 9, NA, NA)
median_possible_values(x1)
#> [1] 7.0 7.5 8.5 9.0
median_range(x1)
#> [1] 7 9

# The output of both functions
# is often the same:
x2 <- c(7, 7, 7, 8, NA, NA)
median_possible_values(x2)
#> [1] 7.0 7.5
median_range(x2)
#> [1] 7.0 7.5

# If there are missing values at
# the median, even these functions
# return `NA`:
x3 <- c(1, 1, 2, NA, NA, NA)
median_possible_values(x3)
#> [1] NA
median_range(x3)
#> [1] NA NA